Separate endpoints for receiving and sending + config files.

+++ cpp/exe/proxy/tcp/AsioTcpSocketWrapper.h
@@ -59,7 +59,7 @@ public:
             }
             else
             {
-                const auto rx_data = this->rx_buffer.as_rslice().take(num_rx);
+                const auto rx_data = this->rx_buffer.as_rslice().take(static_cast<uint32_t>(num_rx));
                 FORMAT_LOG_BLOCK(this->logger, ssp21::levels::debug, "complete socket rx: %u, tx: %s", rx_data.length(), bool_str(this->is_tx_active));
                 if (this->logger.is_enabled(ssp21::levels::debug))
                 {
